*: better large file size support
now 64-bit file sizes on windows are supported everywhere-ish :)

there are a couple hacks here and there since we no longer have
off_t nor pid_t, sigh. though really, we shouldn't be using
sched_delay() at all; if anything it should be that on POSIX
and simply be mt_msleep(1) ? on other platforms

+++ include/dmoz.h
@@ -83,7 +83,7 @@ struct dmoz_file {
 
 	/*struct stat stat;*/
 	time_t timestamp; /* stat.st_mtime */
-	size_t filesize; /* stat.st_size */
+	uint64_t filesize; /* stat.st_size */
 
 	/* if ((type & TYPE_EXT_DATA_MASK) == 0) nothing below this point will
 	be defined (call dmoz_{fill,filter}_ext_data to define it) */
